requirements.txt
flask
flask-cors
python-dotenv
openai
azure-search-documents==11.4.0b7
azure-storage-blob
